🚚  ProjectCard stlye 분리
